requirements.txt
diffusers==0.29.2
h5py==3.11.0
matplotlib==3.7.5
numpy==1.24.3
pandas==2.0.3
Pillow==11.0.0
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.14.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
wandb==0.18.7
